Fixes to model class
nn.Module inherits from dataclass, so there is no need to explicitly decorate it.
The FullyConnectedModule's call implementation created an extra layer after the input with the same size as the expected input layer. The correct behavior is to simply use the input size when initiating the params.
Added input and output sizes to abstract base class.