Replacing fors with scans.
Added to datasets method for returning a scannable version of the dataset.
Reinplamented batch_training using scan, and removed old relevent functions.
Some updates to docstrings.
Commented out tests of deleted methods (the tests are not yet deleted in case they can be reused for tests of other similar functions).