Jitting training.
Rewrote `jitted_training_epoch()` for jitted training to be compatible with scanning.
Change the handling of history in `FullyConnectedEstimator`. It is now stored as an array which can be converted using the method `FullyConnectedEstimator.history_description()`.
`jitted_training()` accepts `epochs` and `metrics` as array to better facilitate jitting.
Updated existing docstring.
Added docstrings to new methods (some of these are partially incomplete, since the methods are still under development).
Minor code clean-up based on these changes.