Dataset randomness.
Added an option to generate an empty dataset based on an existing one.
Added an option to change the seed of a dataset via a property.
Minor fixes to typehints.
When fitting an estimator, the generated test/ validation datasets are now generated with separate rng keys.
If validation_size is None if estimator, an empty dataset is now used.